add add menu function finish

+++ McDonald's-Kiosk/src/Menu.java
@@ -3,11 +3,11 @@ import javax.swing.ImageIcon;
 public class Menu {
 		private int id;
         private String name;
-        private double price;
+        private int price;
         private String description;
         private String image;
 
-        Menu(int id, String name, double price, String description, String image) {
+        Menu(int id, String name, int price, String description, String image) {
         	this.id = id;
             this.name = name;
             this.price = price;
@@ -23,7 +23,7 @@ public class Menu {
             return name;
         }
 
-        public double getPrice() {
+        public int getPrice() {
             return price;
         }
